requirements.txt
openai 
langsmith 
promptlayer
fastapi 
uvicorn
python-dotenv
